ONECOND-736 - Retry option in handler Action

+++ core/src/main/java/com/netflix/conductor/core/execution/WorkflowExecutor.java
@@ -258,6 +258,8 @@ public class WorkflowExecutor {
 
 		// Get the workflow
 		Workflow workflow = edao.getWorkflow(workflowId);
+		if (workflow == null)
+			throw new ApplicationException(Code.NOT_FOUND, "No workflow found with id " + workflowId);
 
 		// If the task Id is null it implies that the entire workflow has to be rerun
 		if( taskId == null) {
@@ -403,6 +405,9 @@ public class WorkflowExecutor {
 
 	public void rewind(String workflowId, String correlationId) throws Exception {
 		Workflow workflow = edao.getWorkflow(workflowId, true);
+		if (workflow == null)
+			throw new ApplicationException(Code.NOT_FOUND, "No workflow found with id " + workflowId);
+
 		if (!workflow.getStatus().isTerminal()) {
 			logger.error("Workflow is still running. status=" + workflow.getStatus()+",workflowId="+workflow.getWorkflowId()+",correlationId="+workflow.getCorrelationId()+ ",contextUser=" + workflow.getContextUser());
 			throw new ApplicationException(Code.CONFLICT, "Workflow is still running. status=" + workflow.getStatus());
@@ -433,6 +438,9 @@ public class WorkflowExecutor {
 
 	public void retry(String workflowId, String correlationId) throws Exception {
 		Workflow workflow = edao.getWorkflow(workflowId, true);
+		if (workflow == null)
+			throw new ApplicationException(Code.NOT_FOUND, "No workflow found with id " + workflowId);
+
 		if (!workflow.getStatus().isTerminal()) {
 			logger.error("Workflow is still running. status=" + workflow.getStatus()+",workflowId="+workflow.getWorkflowId()+",correlationId="+workflow.getCorrelationId() + ",contextUser=" + workflow.getContextUser());
 			throw new ApplicationException(Code.CONFLICT, "Workflow is still running.  status=" + workflow.getStatus());
@@ -595,8 +603,45 @@ public class WorkflowExecutor {
 		logger.debug("Workflow has completed, workflowId=" + wf.getWorkflowId() + ",correlationId=" + wf.getCorrelationId() + ",contextUser=" + workflow.getContextUser());
 	}
 
+	public void forceCompleteWorkflow(String workflowId) throws Exception {
+		Workflow workflow = edao.getWorkflow(workflowId, true);
+		if (workflow == null)
+			throw new ApplicationException(Code.NOT_FOUND, "No workflow found with id " + workflowId);
+
+		if (workflow.getStatus() == WorkflowStatus.COMPLETED) {
+			throw new ApplicationException(Code.CONFLICT, "Workflow has already been completed");
+		}
+
+		workflow.setStatus(WorkflowStatus.COMPLETED);
+		edao.updateWorkflow(workflow);
+		logger.debug("Workflow is force completed. workflowId=" + workflowId + ",correlationId=" + workflow.getCorrelationId()
+			+ ",contextUser=" + workflow.getContextUser());
+		cancelTasks(workflow, workflow.getTasks());
+
+		// If the following lines, for some reason fails, the sweep will take
+		// care of this again!
+		if (workflow.getParentWorkflowId() != null) {
+			Workflow parent = edao.getWorkflow(workflow.getParentWorkflowId(), false);
+			decide(parent.getWorkflowId());
+		}
+
+		//remove from the sweep queue
+		queue.remove(deciderQueue, workflowId);
+
+		// metrics
+		Monitors.recordWorkflowCancel(workflow);
+
+		// send wf end message
+		notifyWorkflowStatus(workflow, StartEndState.end);
+
+		logger.debug("Workflow has force completed, workflowId=" + workflow.getWorkflowId()+",correlationId="+workflow.getCorrelationId()+",contextUser=" + workflow.getContextUser());
+	}
+
 	public String cancelWorkflow(String workflowId , String reason) throws Exception {
 		Workflow workflow = edao.getWorkflow(workflowId, true);
+		if (workflow == null)
+			throw new ApplicationException(Code.NOT_FOUND, "No workflow found with id " + workflowId);
+
 		if (workflow.getStatus().isTerminal() || workflow.getStatus().equals(WorkflowStatus.CANCELLED) ) {
 			String msg = "Workflow can not be cancelled because its already "+workflow.getStatus() ;
 			logger.error("Workflow can not be cancelled because its already " + workflow.getStatus()+",workflowId="+workflow.getWorkflowId()+",correlationId="+workflow.getCorrelationId()+",contextUser=" + workflow.getContextUser());
@@ -675,6 +720,8 @@ public class WorkflowExecutor {
 
 	public String reset(String workflowId, String reason) throws Exception {
 		Workflow workflow = edao.getWorkflow(workflowId, true);
+		if (workflow == null)
+			throw new ApplicationException(Code.NOT_FOUND, "No workflow found with id " + workflowId);
 
 		if (!workflow.getStatus().isTerminal()) {
 			workflow.setStatus(WorkflowStatus.RESET);
@@ -1069,6 +1116,9 @@ public class WorkflowExecutor {
 	public void pauseWorkflow(String workflowId,String correlationId) throws Exception {
 		WorkflowStatus status = WorkflowStatus.PAUSED;
 		Workflow workflow = edao.getWorkflow(workflowId, false);
+		if (workflow == null)
+			throw new ApplicationException(Code.NOT_FOUND, "No workflow found with id " + workflowId);
+
 		if(workflow.getStatus().isTerminal()){
 			throw new ApplicationException(Code.CONFLICT, "Workflow id " + workflowId + " has ended, status cannot be updated.");
 		}
@@ -1087,6 +1137,9 @@ public class WorkflowExecutor {
 
 	public void resumeWorkflow(String workflowId,String correlationId) throws Exception{
 		Workflow workflow = edao.getWorkflow(workflowId, false);
+		if (workflow == null)
+			throw new ApplicationException(Code.NOT_FOUND, "No workflow found with id " + workflowId);
+
 		if(!workflow.getStatus().equals(WorkflowStatus.PAUSED)){
 			logger.error("Workflow is not is not PAUSED so cannot resume. Current status=" + workflow.getStatus() + ",workflowId=" + workflow.getWorkflowId()+",correlationId=" + workflow.getCorrelationId() + ",contextUser=" + workflow.getContextUser());
 			throw new IllegalStateException("The workflow " + workflowId + " is not is not PAUSED so cannot resume");
@@ -1159,6 +1212,9 @@ public class WorkflowExecutor {
 
 	public void removeWorkflow(String workflowId) {
 		Workflow workflow = getWorkflow(workflowId, false);
+		if (workflow == null)
+			throw new ApplicationException(Code.NOT_FOUND, "No workflow found with id " + workflowId);
+
 		edao.removeWorkflow(workflowId);
 		// metrics
 		Monitors.recordWorkflowRemove(workflow);
@@ -1316,8 +1372,6 @@ public class WorkflowExecutor {
 
 	@VisibleForTesting
 	boolean scheduleTask(Workflow workflow, List<Task> tasks) throws Exception {
-		logger.debug("Schedule tasks/input " + tasks.stream().map(Task::getReferenceTaskName).collect(Collectors.toList()) + ",workflowId=" + workflow.getWorkflowId() + ",correlationId=" + workflow.getCorrelationId() + ",contextUser=" + workflow.getContextUser());
-
 		if (tasks == null || tasks.isEmpty()) {
 			return false;
 		}
@@ -1328,13 +1382,8 @@ public class WorkflowExecutor {
 		}
 
 		List<Task> created = edao.createTasks(tasks);
-		logger.debug("Schedule tasks/created " + created.stream().map(Task::getReferenceTaskName).collect(Collectors.toList()) + ",workflowId=" + workflow.getWorkflowId() + ",correlationId=" + workflow.getCorrelationId() + ",contextUser=" + workflow.getContextUser());
-
 		List<Task> createdSystemTasks = created.stream().filter(task -> SystemTaskType.is(task.getTaskType())).collect(Collectors.toList());
-		logger.debug("Schedule tasks/created system " + createdSystemTasks.stream().map(Task::getReferenceTaskName).collect(Collectors.toList()) + ",workflowId=" + workflow.getWorkflowId() + ",correlationId=" + workflow.getCorrelationId() + ",contextUser=" + workflow.getContextUser());
-
 		List<Task> toBeQueued = created.stream().filter(task -> !SystemTaskType.is(task.getTaskType())).collect(Collectors.toList());
-
 		boolean startedSystemTasks = false;
 		for(Task task : createdSystemTasks) {
 
@@ -1344,7 +1393,6 @@ public class WorkflowExecutor {
 			}
 			task.setStartTime(System.currentTimeMillis());
 			if(!stt.isAsync()) {
-				logger.debug("Schedule tasks/starting " + task + ",workflowId=" + workflow.getWorkflowId() + ",correlationId=" + workflow.getCorrelationId() + ",contextUser=" + workflow.getContextUser());
 				notifyTaskStatus(task, StartEndState.start);
 				stt.start(workflow, task, this);
 				startedSystemTasks = true;
@@ -1357,7 +1405,6 @@ public class WorkflowExecutor {
 			}
 		}
 
-		logger.debug("Schedule tasks/toBeQueued " + toBeQueued.stream().map(Task::getReferenceTaskName).collect(Collectors.toList()) + ",workflowId=" + workflow.getWorkflowId() + ",correlationId=" + workflow.getCorrelationId() + ",contextUser=" + workflow.getContextUser());
 		addTaskToQueue(toBeQueued);
 		return startedSystemTasks;
 	}
@@ -1624,6 +1671,11 @@ public class WorkflowExecutor {
 
 		// Get the workflow and find related task
 		Workflow workflow = getWorkflow(workflowId, true);
+		if (workflow == null) {
+			logger.debug("No workflow found with id " + workflowId);
+			return;
+		}
+
 		workflow.setStatus(WorkflowStatus.RUNNING);
 		workflow.setReasonForIncompletion(null);
 		edao.updateWorkflow(workflow);
